requirements.txt
pydantic
python-dotenv
openai
fastapi
uvicorn